requirements.txt
pydantic~=1.10.4
requests~=2.28.1
pystache~=0.6.0
beautifulsoup4
uvicorn
fastapi
PyJWT~=2.6.0
python-dotenv~=1.0.0
cryptography